Move WidgetRect(s) to own file
